added password validation
